Reusing filterQueryAdapter to build sortFilter
