+++ src/main/java/com/vivareal/search/api/model/query/Sort.java
@@ -1,14 +1,11 @@
 package com.vivareal.search.api.model.query;
 
 import com.google.common.base.Objects;
-import org.apache.commons.collections.CollectionUtils;
 
 import java.util.*;
 
 import static com.google.common.base.Objects.equal;
 import static java.lang.String.format;
-import static java.util.Objects.isNull;
-import static java.util.Optional.ofNullable;
 import static java.util.stream.Collectors.joining;
 import static org.apache.commons.lang.StringUtils.EMPTY;
 
@@ -42,24 +39,13 @@ public final class Sort extends AbstractSet<Sort.Item> {
     public static class Item {
         private final Field field;
         private final OrderOperator orderOperator;
-        private Optional<QueryFragmentList> queryFragmentList;
+        private Optional<QueryFragment> queryFragment;
 
         private Item(Field field, OrderOperator orderOperator, Optional<QueryFragment> queryFragment) {
             this.field = field;
             this.orderOperator = orderOperator;
-            this.queryFragmentList = fromQueryFragment(queryFragment);
+            this.queryFragment = queryFragment;
         }
-
-        private Optional<QueryFragmentList> fromQueryFragment(Optional<QueryFragment> queryFragment) {
-            return queryFragment.filter(qf -> qf instanceof QueryFragmentList)
-                .map(qf -> (QueryFragmentList) qf)
-                .filter(CollectionUtils::isNotEmpty)
-
-                // FIX ME - We are waiting a refactor for ElasticSearchQueryAdapter in order to reuse the Filter created from a QueryFragment
-                .filter(qfl -> qfl.get(0) instanceof QueryFragmentItem)
-                ;
-        }
-
 
         public Field getField() {
             return field;
@@ -69,13 +55,13 @@ public final class Sort extends AbstractSet<Sort.Item> {
             return orderOperator;
         }
 
-        public Optional<QueryFragmentList> getQueryFragmentList() {
-            return queryFragmentList;
+        public Optional<QueryFragment> getQueryFragment() {
+            return queryFragment;
         }
 
         @Override
         public String toString() {
-            return format("%s %s %s", field, orderOperator, queryFragmentList.map(QueryFragment::toString).orElse(EMPTY)).trim();
+            return format("%s %s %s", field, orderOperator, queryFragment.map(QueryFragment::toString).orElse(EMPTY)).trim();
         }
 
         @Override
@@ -85,12 +71,12 @@ public final class Sort extends AbstractSet<Sort.Item> {
             Item item = (Item) o;
             return equal(field, item.field)
                 && equal(orderOperator, item.orderOperator)
-                && equal(queryFragmentList, item.queryFragmentList);
+                && equal(queryFragment, item.queryFragment);
         }
 
         @Override
         public int hashCode() {
-            return Objects.hashCode(field, orderOperator, queryFragmentList);
+            return Objects.hashCode(field, orderOperator, queryFragment);
         }
     }
 }
