+++ src/test/java/com/vivareal/search/api/adapter/SortQueryAdapterTest.java
@@ -3,6 +3,7 @@ package com.vivareal.search.api.adapter;
 import com.vivareal.search.api.model.http.SearchApiRequest;
 import com.vivareal.search.api.model.parser.*;
 import org.elasticsearch.action.search.SearchRequestBuilder;
+import org.elasticsearch.index.query.BoolQueryBuilder;
 import org.elasticsearch.index.query.TermsQueryBuilder;
 import org.elasticsearch.search.sort.FieldSortBuilder;
 import org.elasticsearch.search.sort.SortOrder;
@@ -15,17 +16,21 @@ import java.util.Map;
 import static com.vivareal.search.api.configuration.environment.RemoteProperties.ES_DEFAULT_SORT;
 import static com.vivareal.search.api.model.http.SearchApiRequestBuilder.INDEX_NAME;
 import static com.vivareal.search.api.model.mapping.MappingType.FIELD_TYPE_NESTED;
+import static org.elasticsearch.index.query.QueryBuilders.boolQuery;
 import static org.elasticsearch.index.query.QueryBuilders.termQuery;
 import static org.elasticsearch.search.sort.SortOrder.ASC;
 import static org.elasticsearch.search.sort.SortOrder.DESC;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNull;
+import static org.mockito.Matchers.any;
+import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 public class SortQueryAdapterTest extends SearchTransportClientMock {
 
     private SettingsAdapter<Map<String, Map<String, Object>>, String> settingsAdapter;
+    private FilterQueryAdapter filterQueryAdapter;
 
     private SortQueryAdapter sortQueryAdapter;
 
@@ -39,7 +44,8 @@ public class SortQueryAdapterTest extends SearchTransportClientMock {
         SortParser sortParser = new SortParser(fieldParser, operatorParser, queryParser);
 
         this.settingsAdapter = mock(SettingsAdapter.class);
-        this.sortQueryAdapter = new SortQueryAdapter(settingsAdapter, sortParser);
+        this.filterQueryAdapter = mock(FilterQueryAdapter.class);
+        this.sortQueryAdapter = new SortQueryAdapter(settingsAdapter, sortParser, filterQueryAdapter);
     }
 
     @BeforeClass
@@ -101,7 +107,6 @@ public class SortQueryAdapterTest extends SearchTransportClientMock {
         assertNull(sorts.get(2).getNestedPath());
     }
 
-
     @Test
     public void shouldApplySortFilterWhenExplicit() {
         String fieldName1 = "field";
@@ -111,13 +116,14 @@ public class SortQueryAdapterTest extends SearchTransportClientMock {
         SortOrder sortOrder2 = DESC;
         String sortFilter2 = "sortFilter: fieldName EQ \"value\"";
 
-
         SearchRequestBuilder requestBuilder = transportClient.prepareSearch(INDEX_NAME);
         SearchApiRequest request = fullRequest.build();
         request.setSort(fieldName1 + " " + sortOrder1.name() + ", " + fieldName2 + " " + sortOrder2.name() + " " + sortFilter2);
 
         when(settingsAdapter.isTypeOf(request.getIndex(), fieldName1, FIELD_TYPE_NESTED)).thenReturn(false);
         when(settingsAdapter.isTypeOf(request.getIndex(), fieldName2.split("\\.")[0], FIELD_TYPE_NESTED)).thenReturn(true);
+
+        BoolQueryBuilder boolQueryBuilder = boolQuery();
 
         sortQueryAdapter.apply(requestBuilder, request);
         List<FieldSortBuilder> sorts = (List) requestBuilder.request().source().sorts();
@@ -129,7 +135,7 @@ public class SortQueryAdapterTest extends SearchTransportClientMock {
         assertEquals(fieldName2, sorts.get(1).getFieldName());
         assertEquals(sortOrder2, sorts.get(1).order());
         assertEquals("nested", sorts.get(1).getNestedPath());
-        assertEquals(termQuery("fieldName", "value"), sorts.get(1).getNestedFilter());
+        assertEquals(boolQueryBuilder, sorts.get(1).getNestedFilter());
 
         assertEquals("_uid", sorts.get(2).getFieldName());
         assertEquals("DESC", sorts.get(2).order().name());
